:bug: fix: always allocate the max needed memory
The way the code looked before, there could be a string size overflow if the input phone was too large. For example, if it had 14 digits.

The way it is now, the result variable will always be allocated with the size of the input phone, so there is no possibility of the overflow that existed before.